+++ R/model-tracking.R
@@ -1,11 +1,21 @@
 
-#' Logs model performans data.frame to BigQuery table
+#' Logs model performance data.frame to BigQuery table
 #'
 #' @export
 #' @param data data with metrics of the model
 logModelPerformance <- function(data) {
 
-  data <- data[, c("project", "model", "metric", "group", "size", "value", "dataset")]
+
+  data <- data[,
+               c(
+                 "project",
+                 "model",
+                 "metric",
+                 "group",
+                 "size",
+                 "value",
+                 "dataset"
+                )]
   data <- cbind(date = Sys.time(), data)
 
   job <- insert_upload_job(project = Sys.getenv("BIGQUERY_PROJECT"),