+++ tests/testthat/test.bq.table.R
@@ -17,7 +17,7 @@ test_that("table is created from schema file", {
     list(name = "first_name",
          type = "STRING")
   )
-  expected.partition = TRUE
+  expected.partition <- TRUE
 })
 
 test_that("correct schema is generated for a given dataset", {
@@ -26,4 +26,10 @@ test_that("correct schema is generated for a given dataset", {
   expect_equal(res$name, c("speed", "dist"))
   expect_equal(res$type, c("FLOAT", "FLOAT"))
   expect_equal(res$mode, c("NULLABLE", "NULLABLE"))
+})
+
+test_that("Correct extension is assigned to a given file format", {
+  res <- extensionFromFormat("CSV", "GZIP")
+  expect <- "csv.gz"
+  expect_equal(res, expect)
 })